# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IKODI Student Sponsorship Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all system users
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [secretary, admin, super_admin]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new system user (Super Admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserWithCredentials"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    put:
      operationId: updateUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted
...
  /students:
    get:
      operationId: listStudents
      tags: [students]
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: schoolId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, graduated]
        - name: sponsorshipStatus
          in: query
          schema:
            type: string
            enum: [sponsored, partial, unsponsored]
        - name: term
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Student"
    post:
      operationId: createStudent
      tags: [students]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
  /students/{id}:
    get:
      operationId: getStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Student details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentDetail"
    put:
      operationId: updateStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentBody"
      responses:
        "200":
          description: Student updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Student"
    delete:
      operationId: deleteStudent
      tags: [students]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Student deleted
  /students/{id}/academic-records:
    get:
      operationId: getStudentAcademicRecords
      tags: [academic]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Academic records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AcademicRecord"
    post:
      operationId: addAcademicRecord
      tags: [academic]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAcademicRecordBody"
      responses:
        "201":
          description: Academic record added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcademicRecord"
...
  /reports/students:
    get:
      operationId: getStudentsReport
      tags: [reports]
      parameters:
        - name: term
          in: query
          schema:
            type: string
        - name: schoolCategory
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Students report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentsReport"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [secretary, admin, super_admin]
        isActive:
          type: boolean
        lastLogin:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - username
        - fullName
        - role
        - isActive
        - createdAt
    UserWithCredentials:
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          properties:
            generatedPassword:
              type: string
          required:
            - generatedPassword
    CreateUserBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [secretary, admin, super_admin]
      required:
        - fullName
        - role
    UpdateUserBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [secretary, admin, super_admin]
        isActive:
          type: boolean
...
    Student:
      type: object
      properties:
        id:
          type: integer
        admissionNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        schoolId:
          type: integer
          nullable: true
        schoolName:
          type: string
          nullable: true
        schoolCategory:
          type: string
          nullable: true
        course:
          type: string
          nullable: true
        currentLevel:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, graduated]
        sponsorshipStatus:
          type: string
          enum: [sponsored, partial, unsponsored]
        totalFees:
          type: number
          nullable: true
        paidAmount:
          type: number
          nullable: true
        balance:
          type: number
          nullable: true
        currentTerm:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - admissionNumber
        - firstName
        - lastName
        - status
        - sponsorshipStatus
        - createdAt
    StudentDetail:
      allOf:
        - $ref: "#/components/schemas/Student"
        - type: object
          properties:
            guardians:
              type: array
              items:
                $ref: "#/components/schemas/Guardian"
            sponsorships:
              type: array
              items:
                $ref: "#/components/schemas/Sponsorship"
            recentPayments:
              type: array
              items:
                $ref: "#/components/schemas/Payment"
            academicRecords:
              type: array
              items:
                $ref: "#/components/schemas/AcademicRecord"
    CreateStudentBody:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          nullable: true
        gender:
          type: string
          enum: [male, female, other]
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        schoolId:
          type: integer
          nullable: true
        course:
          type: string
          nullable: true
        currentLevel:
          type: string
          nullable: true
        currentTerm:
          type: string
          nullable: true
        totalFees:
          type: number
          nullable: true
        status:
          type: string
          enum: [active, inactive, graduated]
        guardianName:
          type: string
          nullable: true
        guardianRelationship:
          type: string
          nullable: true
        guardianPhone:
          type: string
          nullable: true
        guardianEmail:
          type: string
          nullable: true
      required:
        - firstName
        - lastName
    Guardian:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        name:
          type: string
        relationship:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
      required:
        - id
        - studentId
        - name
    AcademicRecord:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        term:
          type: string
        year:
          type: integer
        subject:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        gpa:
          type: number
          nullable: true
        remarks:
          type: string
          nullable: true
        reportCardUrl:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - studentId
        - term
        - year
        - createdAt
    CreateAcademicRecordBody:
      type: object
      properties:
        term:
          type: string
        year:
          type: integer
        subject:
          type: string
          nullable: true
        grade:
          type: string
          nullable: true
        gpa:
          type: number
          nullable: true
        remarks:
          type: string
          nullable: true
      required:
        - term
        - year
...
    Sponsorship:
      type: object
      properties:
        id:
          type: integer
        studentId:
          type: integer
        sponsorId:
          type: integer
        studentName:
          type: string
          nullable: true
        sponsorName:
          type: string
          nullable: true
        coverageType:
          type: string
          enum: [full, partial]
        amount:
          type: number
        startDate:
          type: string
        endDate:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, inactive, completed]
        term:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - studentId
        - sponsorId
        - coverageType
        - amount
        - startDate
        - status
        - createdAt
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        sponsorshipId:
          type: integer
          nullable: true
        sponsorId:
          type: integer
          nullable: true
        studentId:
          type: integer
          nullable: true
        studentName:
          type: string
          nullable: true
        sponsorName:
          type: string
          nullable: true
        amount:
          type: number
        paymentDate:
          type: string
        paymentMethod:
          type: string
          enum: [mpesa, bank_transfer, cash, cheque, online]
        referenceNumber:
          type: string
          nullable: true
        purpose:
          type: string
          nullable: true
        term:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - amount
        - paymentDate
        - paymentMethod
        - createdAt
...
    StudentsReport:
      type: object
      properties:
        term:
          type: string
          nullable: true
        totalStudents:
          type: integer
        bySchoolCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required:
              - category
              - count
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
      required:
        - totalStudents
        - bySchoolCategory
        - students
...