# lib/api-zod/src/generated/types/createAcademicRecordBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * IKODI Student Sponsorship Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAcademicRecordBody {
  term: string;
  year: number;
  subject?: string | null;
  grade?: string | null;
  gpa?: number | null;
  remarks?: string | null;
}